refactor(RacingCar) : 자동차 경추 리팩토링(5단계)
- 패키지 분리
