+++ src/main/java/study/racingcar/domain/RacingGame.java
@@ -1,4 +1,4 @@
-package study.racingcar;
+package study.racingcar.domain;
 
 import java.util.List;
