+++ src/main/java/study/racingcar/domain/SettingGame.java
@@ -1,4 +1,6 @@
-package study.racingcar;
+package study.racingcar.domain;
+
+import study.racingcar.strategy.CarMoveStrategy;
 
 import java.util.ArrayList;
 import java.util.List;
