+++ src/main/java/study/racingcar/strategy/MoveStrategy.java
@@ -1,4 +1,4 @@
-package study.racingcar;
+package study.racingcar.strategy;
 
 public interface MoveStrategy {
