+++ src/main/java/study/racingcar/view/ResultView.java
@@ -1,4 +1,6 @@
-package study.racingcar;
+package study.racingcar.view;
+
+import study.racingcar.domain.Car;
 
 import java.util.List;
