+++ src/test/java/study/racingcar/CarMoveStrategyTest.java
@@ -2,6 +2,7 @@ package study.racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import study.racingcar.strategy.CarMoveStrategy;
 
 import static org.assertj.core.api.Assertions.assertThat;
