+++ src/test/java/study/racingcar/CarTest.java
@@ -3,6 +3,8 @@ package study.racingcar;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import study.racingcar.domain.Car;
+import study.racingcar.strategy.CarMoveStrategy;
 
 import java.util.InputMismatchException;
