+++ src/test/java/study/racingcar/RacingGameTest.java
@@ -3,6 +3,8 @@ package study.racingcar;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import study.racingcar.domain.RacingGame;
+import study.racingcar.domain.RacingGameInputs;
 
 import static org.assertj.core.api.Assertions.assertThat;
